Enhance data service error handling and code organization
• Add HTTP status text to all error messages for better debugging
• Wrap JSON parsing in try/catch with descriptive error messages
• Add comprehensive caching and error handling test coverage
• Externalize COURT_MAPPINGS to separate constants file
• Test malformed JSON, network errors, and fetch rejections
