+++ src/types/court.types.ts
@@ -2,98 +2,3 @@ export interface Court {
   code: string;
   name: string;
 }
-
-export const COURT_MAPPINGS: Record<string, string> = {
-  almd: 'Alabama Middle District',
-  alnd: 'Alabama Northern District',
-  alsd: 'Alabama Southern District',
-  akd: 'Alaska District',
-  azd: 'Arizona District',
-  ared: 'Arkansas Eastern District',
-  arwd: 'Arkansas Western District',
-  cacd: 'California Central District',
-  caed: 'California Eastern District',
-  cand: 'California Northern District',
-  casd: 'California Southern District',
-  cod: 'Colorado District',
-  ctd: 'Connecticut District',
-  ded: 'Delaware District',
-  dcd: 'District of Columbia',
-  flmd: 'Florida Middle District',
-  flnd: 'Florida Northern District',
-  flsd: 'Florida Southern District',
-  gad: 'Georgia District',
-  gamd: 'Georgia Middle District',
-  gand: 'Georgia Northern District',
-  gasd: 'Georgia Southern District',
-  hid: 'Hawaii District',
-  idd: 'Idaho District',
-  ilcd: 'Illinois Central District',
-  ilnd: 'Illinois Northern District',
-  ilsd: 'Illinois Southern District',
-  innd: 'Indiana Northern District',
-  insd: 'Indiana Southern District',
-  iasd: 'Iowa Southern District',
-  iand: 'Iowa Northern District',
-  ksd: 'Kansas District',
-  kyed: 'Kentucky Eastern District',
-  kywd: 'Kentucky Western District',
-  laed: 'Louisiana Eastern District',
-  lamd: 'Louisiana Middle District',
-  lawd: 'Louisiana Western District',
-  med: 'Maine District',
-  mdd: 'Maryland District',
-  mad: 'Massachusetts District',
-  mied: 'Michigan Eastern District',
-  miwd: 'Michigan Western District',
-  mnd: 'Minnesota District',
-  msnd: 'Mississippi Northern District',
-  mssd: 'Mississippi Southern District',
-  moed: 'Missouri Eastern District',
-  mowd: 'Missouri Western District',
-  mtd: 'Montana District',
-  ned: 'Nebraska District',
-  nvd: 'Nevada District',
-  nhd: 'New Hampshire District',
-  njd: 'New Jersey District',
-  nmd: 'New Mexico District',
-  nyed: 'New York Eastern District',
-  nynd: 'New York Northern District',
-  nysd: 'New York Southern District',
-  nywd: 'New York Western District',
-  nysb: 'New York Southern Bankruptcy',
-  ncmd: 'North Carolina Middle District',
-  nced: 'North Carolina Eastern District',
-  ncwd: 'North Carolina Western District',
-  ndd: 'North Dakota District',
-  ohnd: 'Ohio Northern District',
-  ohsd: 'Ohio Southern District',
-  oked: 'Oklahoma Eastern District',
-  oknd: 'Oklahoma Northern District',
-  okwd: 'Oklahoma Western District',
-  ord: 'Oregon District',
-  paed: 'Pennsylvania Eastern District',
-  pamd: 'Pennsylvania Middle District',
-  pawd: 'Pennsylvania Western District',
-  rid: 'Rhode Island District',
-  scd: 'South Carolina District',
-  sdd: 'South Dakota District',
-  tned: 'Tennessee Eastern District',
-  tnmd: 'Tennessee Middle District',
-  tnwd: 'Tennessee Western District',
-  txed: 'Texas Eastern District',
-  txnd: 'Texas Northern District',
-  txsd: 'Texas Southern District',
-  txwd: 'Texas Western District',
-  utd: 'Utah District',
-  vtd: 'Vermont District',
-  vaed: 'Virginia Eastern District',
-  vawd: 'Virginia Western District',
-  waed: 'Washington Eastern District',
-  wawd: 'Washington Western District',
-  wvnd: 'West Virginia Northern District',
-  wvsd: 'West Virginia Southern District',
-  wied: 'Wisconsin Eastern District',
-  wiwd: 'Wisconsin Western District',
-  wyd: 'Wyoming District',
-};
